[Silver IV] Title: 스케이트 연습, Time: 392 ms, Memory: 68740 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/28324. 스케이트 연습/스케이트 연습.java
@@ -14,7 +14,7 @@ class Main {
             arr[i] = Integer.parseInt(st.nextToken());
         }
 
-        int sum = 0;
+        long sum = 0L;
         int curSpeed = 0;
 
         for (int i = N ; i >= 1 ; i--) {
